chore: generate code

+++ modules/overrungl.core/src/main/java/overrungl/internal/RuntimeHelper.java
@@ -21,6 +21,7 @@ import io.github.overrun.platform.Platform;
 import overrungl.OverrunGLConfigurations;
 import overrungl.OverrunGL;
 
+import java.io.File;
 import java.io.IOException;
 import java.lang.foreign.*;
 import java.lang.invoke.MethodHandle;
@@ -85,37 +86,49 @@ public final class RuntimeHelper {
         Path uri;
         // 1. Load from natives directory
         var localFile = Path.of(System.getProperty("overrungl.natives", "."), path);
-        var localJLPFile = Path.of(System.getProperty("java.library.path"), path);
         if (Files.exists(localFile)) {
             uri = localFile;
-        } else if (Files.exists(localJLPFile)) {
-            uri = localJLPFile;
         } else {
-            // 2. Load from classpath
-            try {
-                if (!Files.exists(tmpdir)) {
-                    // Create directory
-                    Files.createDirectories(tmpdir);
-                } else if (!Files.isDirectory(tmpdir)) {
-                    // Remove
-                    Files.delete(tmpdir);
-                    // Create directory
-                    Files.createDirectories(tmpdir);
-                }
-            } catch (IOException e) {
-                throw new IllegalStateException("Couldn't create directory: " + tmpdir + "; try setting -Doverrungl.natives or -Djava.library.path to a valid path", e);
-            }
-            var libFile = tmpdir.resolve(basename + "-" + version + suffix);
-            if (!Files.exists(libFile)) {
-                // Extract
-                final String fromPath = "overrungl." + module + "/" + os.familyName() + "-" + Architecture.current() + "/" + path;
-                try (var is = ClassLoader.getSystemResourceAsStream(fromPath)) {
-                    Files.copy(Objects.requireNonNull(is, "File not found in classpath: " + fromPath), libFile);
-                } catch (Exception e) {
-                    throw new IllegalStateException("Couldn't load file: " + libFile.toAbsolutePath().normalize() + " or " + localFile.toAbsolutePath().normalize() + "; try setting -Doverrungl.natives or -Djava.library.path to a valid path", e);
+            // 2. Load from java.library.path
+            boolean found = false;
+            Path jlpPath = null;
+            for (String s : System.getProperty("java.library.path").split(File.pathSeparator)) {
+                var path1 = Path.of(s, path);
+                if (Files.exists(path1)) {
+                    jlpPath = path1;
+                    found = true;
+                    break;
                 }
             }
-            uri = libFile;
+            if (found) {
+                uri = jlpPath;
+            } else {
+                // 3. Load from classpath
+                try {
+                    if (!Files.exists(tmpdir)) {
+                        // Create directory
+                        Files.createDirectories(tmpdir);
+                    } else if (!Files.isDirectory(tmpdir)) {
+                        // Remove
+                        Files.delete(tmpdir);
+                        // Create directory
+                        Files.createDirectories(tmpdir);
+                    }
+                } catch (IOException e) {
+                    throw new IllegalStateException("Couldn't create directory: " + tmpdir + "; try setting -Doverrungl.natives or -Djava.library.path to a valid path", e);
+                }
+                var libFile = tmpdir.resolve(basename + "-" + version + suffix);
+                if (!Files.exists(libFile)) {
+                    // Extract
+                    final String fromPath = "overrungl." + module + "/" + os.familyName() + "-" + Architecture.current() + "/" + path;
+                    try (var is = ClassLoader.getSystemResourceAsStream(fromPath)) {
+                        Files.copy(Objects.requireNonNull(is, "File not found in classpath: " + fromPath), libFile);
+                    } catch (Exception e) {
+                        throw new IllegalStateException("Couldn't load file: " + libFile.toAbsolutePath().normalize() + " or " + localFile.toAbsolutePath().normalize() + "; try setting -Doverrungl.natives or -Djava.library.path to a valid path", e);
+                    }
+                }
+                uri = libFile;
+            }
         }
         if (OverrunGLConfigurations.DEBUG.get()) {
             OverrunGL.apiLog("[OverrunGL] Loading native library from: " + uri);
